+++ fnt/__init__.py
@@ -3,4 +3,4 @@
 This module initializes the fnt package.
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"